refactor: remove data-utils path mappings
