+++ containers/ecr-viewer/src/app/utils/data-utils.tsx
@@ -3,10 +3,6 @@ import { DisplayDataProps } from "@/app/view-data/components/DataDisplay";
 import sanitizeHtml from "sanitize-html";
 import parse from "html-react-parser";
 
-export interface PathMappings {
-  [key: string]: string;
-}
-
 export interface CompleteData {
   availableData: DisplayDataProps[];
   unavailableData: DisplayDataProps[];
